Cleans up studies.
