+++ studies/rjmc/gaussian.py
@@ -2,30 +2,8 @@ import numpy
 import torch
 
 from bayesiantesting.kernels.rjmc import BiasedRJMCSimulation
-from bayesiantesting.models import Model, ModelCollection
-from bayesiantesting.utils import distributions
-
-
-class GaussianModel(Model):
-    """A representation of the two-center Lennard-Jones model, which
-    can be evaluated using a surrogate model against a `NISTDataSet`.
-    """
-
-    def __init__(self, name, prior_settings, loc, scale, weight=1.0):
-
-        super().__init__(name, prior_settings, {})
-
-        self._loc = loc
-        self._scale = scale
-        self._weight = weight
-
-    def evaluate_log_likelihood(self, parameters):
-        return numpy.log(self._weight) + distributions.Normal(
-            self._loc, self._scale
-        ).log_pdf(parameters)
-
-    def compute_percentage_deviations(self, parameters):
-        return {}
+from bayesiantesting.models import ModelCollection
+from bayesiantesting.models.continuous import GaussianModel
 
 
 def main():
