+++ studies/rjmc/run.py
@@ -6,128 +6,12 @@ Created on Thu Oct 31 14:42:37 2019
 @author: owenmadin
 """
 
-import math
-
 import numpy
-import yaml
 
-from bayesiantesting import unit
-from bayesiantesting.datasets.nist import NISTDataSet, NISTDataType
 from bayesiantesting.kernels.rjmc import BiasedRJMCSimulation
-from bayesiantesting.models.continuous import TwoCenterLJModel
 from bayesiantesting.models.discrete import TwoCenterLJModelCollection
-from bayesiantesting.surrogates import StollWerthSurrogate
 from bayesiantesting.utils import distributions
-
-
-def parse_input_yaml(filepath):
-
-    print("Loading simulation params from " + filepath + "...")
-
-    with open(filepath) as file:
-        simulation_params = yaml.load(file, Loader=yaml.SafeLoader)
-
-    return simulation_params
-
-
-def prepare_data(simulation_params):
-    """From input parameters, pull appropriate experimental data and
-    uncertainty information.
-    """
-
-    # Retrieve the constants and thermophysical data
-    data_set = NISTDataSet(simulation_params["compound"])
-
-    # Filter the data to selected conditions.
-    minimum_temperature = (
-        simulation_params["trange"][0]
-        * data_set.critical_temperature.value.to(unit.kelvin).magnitude
-    )
-    maximum_temperature = (
-        simulation_params["trange"][1]
-        * data_set.critical_temperature.value.to(unit.kelvin).magnitude
-    )
-
-    data_set.filter(
-        minimum_temperature * unit.kelvin,
-        maximum_temperature * unit.kelvin,
-        simulation_params["number_data_points"],
-    )
-
-    property_types = []
-
-    if simulation_params["properties"] == "All":
-        property_types.extend(data_set.data_types)
-    else:
-        if "rhol" in simulation_params["properties"]:
-            property_types.append(NISTDataType.LiquidDensity)
-        if "Psat" in simulation_params["properties"]:
-            property_types.append(NISTDataType.SaturationPressure)
-
-    return data_set, property_types
-
-
-def get_model(model_name, data_set, property_types, simulation_params):
-
-    priors = {
-        "epsilon": simulation_params["priors"]["epsilon"],
-        "sigma": simulation_params["priors"]["sigma"],
-    }
-    fixed = {}
-
-    if model_name == "AUA":
-
-        priors["L"] = simulation_params["priors"]["L"]
-        fixed["Q"] = 0.0
-
-    elif model_name == "AUA+Q":
-
-        priors["L"] = simulation_params["priors"]["L"]
-        priors["Q"] = simulation_params["priors"]["Q"]
-
-    elif model_name == "UA":
-
-        fixed["L"] = data_set.bond_length.to(unit.nanometer).magnitude
-        fixed["Q"] = 0.0
-
-    else:
-
-        raise NotImplementedError()
-
-    model = TwoCenterLJModel(
-        model_name,
-        priors,
-        fixed,
-        data_set,
-        property_types,
-        StollWerthSurrogate(data_set.molecular_weight),
-    )
-
-    return model
-
-
-def generate_initial_parameters(model):
-
-    initial_log_p = math.nan
-    initial_parameters = None
-
-    counter = 0
-
-    while math.isnan(initial_log_p) and counter < 1000:
-
-        initial_parameters = model.sample_priors()
-        initial_log_p = model.evaluate_log_posterior(initial_parameters)
-
-        counter += 1
-
-    if numpy.isnan(initial_log_p):
-
-        raise ValueError(
-            "The initial values could not be set without yielding "
-            "a NaN log posterior"
-        )
-
-    return initial_parameters
+from studies.utilities import get_2clj_model, parse_input_yaml, prepare_data
 
 
 def main():
@@ -140,9 +24,9 @@ def main():
 
     # Define the initial parameters as pre-computed MAP values,
     maximum_a_posteriori = [
-        numpy.array([97.0, 0.37850, 0.15]),
-        numpy.array([98.0, 0.37800, 0.15, 0.01]),
-        numpy.array([99.5, 0.37685]),
+        numpy.array([97.08159, 0.37892, 0.14780]),
+        numpy.array([97.29539, 0.37873, 0.14832, 0.01]),
+        numpy.array([99.50692, 0.37684]),
     ]
 
     # Define the mapping distributions
@@ -166,10 +50,13 @@ def main():
 
     # Build the model / models.
     sub_models = [
-        get_model("AUA", data_set, property_types, simulation_params),
-        get_model("AUA+Q", data_set, property_types, simulation_params),
-        get_model("UA", data_set, property_types, simulation_params),
+        get_2clj_model("AUA", data_set, property_types, simulation_params),
+        get_2clj_model("AUA+Q", data_set, property_types, simulation_params),
+        get_2clj_model("UA", data_set, property_types, simulation_params),
     ]
+
+    for mean, model in zip(maximum_a_posteriori, sub_models):
+        print(model.evaluate_log_posterior(mean))
 
     model_collection = TwoCenterLJModelCollection(
         "2CLJ Models", sub_models, mapping_distributions
